#917 refactor: extract shared search foundation
