+++ src/components/search-filters/__tests__/CreateSearchFilterWizard.test.tsx
@@ -46,7 +46,9 @@ vi.mock("@/components/search/filters/PriceSpanFilter.tsx", () => ({
 vi.mock("@/components/search/filters/ProductStateFilter.tsx", () => ({
     ProductStateFilter: () => <div data-testid="state-filter" />,
 }));
-vi.mock("@/components/search/filters/ShopTypeFilter.tsx", () => ({
+vi.mock("@/features/search/common", () => ({
+    SearchFilterFormProvider: ({ children }: { children: React.ReactNode }) => <>{children}</>,
+    SearchQueryProvider: ({ children }: { children: React.ReactNode }) => <>{children}</>,
     ShopTypeFilter: () => <div data-testid="shop-type-filter" />,
 }));
 vi.mock("@/components/search/filters/MerchantFilters.tsx", () => ({
