+++ src/features/search/common/components/SearchBar.tsx
@@ -18,11 +18,11 @@ import { useTranslation } from "react-i18next";
 import type { TFunction } from "i18next";
 import { z } from "zod";
 import { MIN_SEARCH_QUERY_LENGTH } from "@/lib/filterDefaults.ts";
-import { useSearchQueryContext } from "@/hooks/search/useSearchQueryContext.tsx";
+import { useSearchQueryContext } from "@/features/search/common/hooks/useSearchQueryContext.tsx";
 import { useEffect, useMemo, useState } from "react";
 import { toast } from "sonner";
 import { env } from "@/env.ts";
-import { useAnimatedPlaceholder } from "@/hooks/useAnimatedPlaceholder";
+import { useAnimatedPlaceholder } from "@/features/search/common/hooks/useAnimatedPlaceholder.ts";
 import { serializeSearchParams } from "@/lib/searchValidation.ts";
 import type { SearchFilterArguments } from "@/data/internal/search/SearchFilterArguments.ts";
 import { stripLanguageFromPathname } from "@/i18n/routing.ts";
