+++ src/features/search/common/hooks/__tests__/useAnimatedPlaceholder.test.ts
