+++ src/features/search/common/hooks/useAnimatedPlaceholder.ts
